17/08/19 latest changes

+++ RISCV_final_project/mars/mips/instructions/syscalls/SyscallRandDouble.java
@@ -56,20 +56,13 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
           // Input arguments: $a0 = index of pseudorandom number generator
           // Return: $f0 = the next pseudorandom, uniformly distributed double value between 0.0 and 1.0
           // from this random number generator's sequence.
-         Integer index = new Integer(RV32IRegisters.getValue(4));
+         Integer index = new Integer(RV32IRegisters.getValue(4).intValue());
          Random stream = (Random) RandomStreams.randomStreams.get(index);
          if (stream == null) {
             stream = new Random(); // create a non-seeded stream
             RandomStreams.randomStreams.put(index, stream);
-         } 
-       //  try {
+         }
             Coprocessor1.setRegister(0, stream.nextDouble( ));
-        // }     
-       //      catch (InvalidRegisterAccessException e) {   // register ID error in this method
-          //     throw new ProcessingException(statement,
-          //             "Internal error storing double to register (syscall "+this.getNumber()+")",
-          //                   Exceptions.SYSCALL_EXCEPTION);
-       //     }
       }
    
    }
