+++ RISCV_final_project/mars/mips/instructions/syscalls/SyscallRandFloat.java
@@ -56,7 +56,7 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
           // Input arguments: $a0 = index of pseudorandom number generator
           // Return: $f0 = the next pseudorandom, uniformly distributed float value between 0.0 and 1.0
           // from this random number generator's sequence.
-         Integer index = new Integer(RV32IRegisters.getValue(4));
+         Integer index = new Integer(RV32IRegisters.getValue(4).intValue());
          Random stream = (Random) RandomStreams.randomStreams.get(index);
          if (stream == null) {
             stream = new Random(); // create a non-seeded stream
